% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_create_msg_profile}
\alias{tl_create_msg_profile}
\alias{tl_update_msg_profile}
\title{Create or update a messaging profile}
\usage{
tl_create_msg_profile(
  key,
  name = NULL,
  geomatch = TRUE,
  skip_unhealthy = FALSE,
  sticky_sender = TRUE,
  long_code_weight = 1,
  toll_free_weight = 1,
  whitelisted_destinations = "US",
  webhook_url = "",
  webhook_failover_url = "",
  params = NULL,
  ...
)

tl_update_msg_profile(key, profile_id, params)
}
\arguments{
\item{key}{Required. Telnyx API key}

\item{name}{Optional. A user friendly name for the messaging profile.}

\item{geomatch}{Should numbers match area codes? Defaults to TRUE.}

\item{skip_unhealthy}{Should numbers that are blocked not be used? Defaults to FALSE.}

\item{sticky_sender}{Should contacts receive messages from the same number? Defaults to TRUE.}

\item{long_code_weight}{Probability of using Long Code (as a share of the sum of the weights). Defaults to 1.}

\item{toll_free_weight}{Probability of using Toll Free (as a share of the sum of the weights). Defaults to 1.}

\item{whitelisted_destinations}{Vector of country codes numbers can send to. Defaults to 'US'. Should be ISO 3166-1 alpha-2 country codes.}

\item{webhook_url}{Optional. URL for webhook response.}

\item{webhook_failover_url}{Optional. Backup URL for webhook response.}

\item{params}{Optional further arguments to pass into API body. Should be a named list.
Required if using \code{tl_update_msg_profile}, use this to pass in updated values as a named list.}

\item{...}{Unused at this time.}

\item{profile_id}{The messaging profile to update. Required for \code{tl_update_msg_profile}.}
}
\description{
Create a Telnyx Messaging Profile via the Messgaging Profiles endpoint.

See \href{https://developers.telnyx.com/docs/api/v2/messaging/Messaging-Profiles}{here} for more details.
}
