% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_get_messages}
\alias{tl_get_messages}
\title{Get SMS messages.}
\usage{
tl_get_messages(
  start_time,
  end_time = NULL,
  profile_ids = NULL,
  verbose = FALSE,
  v1_token,
  v1_user,
  params = NULL
)
}
\arguments{
\item{start_time}{start time of report. Required. Must include the timezone.}

\item{end_time}{optional time for end of report. Must include the timezone.}

\item{profile_ids}{an optional vector of Messaging Profile ids to filter down to.}

\item{verbose}{defaults to FALSE.}

\item{v1_token}{Telnyx API v1 Token}

\item{v1_user}{Telnyx API v1 user, either an email or the preferred masked API user id.}

\item{params}{optional further arguments to pass into API query. Should be a named list.}
}
\description{
Get SMS messages from the MDR reports endpoint.
This reporting works as an asynchronous call to the report api and returns a CSV parsed into a data table using \code{fread}

See \href{https://developers.telnyx.com/docs/api/v1/reports}{here} for more details.
}
