% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_get_msg_profile_nums}
\alias{tl_get_msg_profile_nums}
\title{List numbers associated with a Messaging Profile}
\usage{
tl_get_msg_profile_nums(key, profile_id)
}
\arguments{
\item{key}{Required. Telnyx API key.}

\item{profile_id}{Required. Messaging Profile id to update all numbers.}
}
\description{
Lists all phone numbers associated with a Messaging Profile and their settings.

See \href{https://developers.telnyx.com/docs/api/v2/messaging/Messaging-Profiles#listMessagingProfilePhoneNumbers}{here} for more details.
}
