% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_order_numbers}
\alias{tl_order_numbers}
\title{Purchase numbers and optionally assign to Messaging Profile.}
\usage{
tl_order_numbers(
  key,
  n,
  search_only = FALSE,
  incl_best_effort = FALSE,
  type = "long-code",
  area_code = NULL,
  city = NULL,
  state = NULL,
  profile_id = NULL,
  params = NULL,
  ...
)
}
\arguments{
\item{key}{Required. Telnyx API key}

\item{n}{Required. Integer number of phone numbers to purchase. Maximum is 100 If less than 100 are available
for the given filter parameters it will purchase all available numbers.}

\item{search_only}{Boolean. If \code{TRUE} will return only the results of the number search and not purchase the numbers. Defaults to \code{FALSE}.}

\item{incl_best_effort}{Boolean. If \code{TRUE} will include 'best effort' results,
i.e. results that are close to the filtered parameters (e.g. geographically) but are not exactly within the filter.
Defaults to \code{FALSE} in order to make this decision explicit by the user.}

\item{type}{the type of phone number to purchase. Options are 'toll-free' and 'long-code'. Defaults to 'long-code'.}

\item{area_code}{Optional. Specify the 3 digit area code the phone number should be associated with as a character string.}

\item{city}{Optional. City that the phone number should be in. It is recommended to use this in conjunction with the \code{state} parameter.}

\item{state}{Optional. Two character state abbreviation that the phone number should be in.}

\item{profile_id}{Optional. The messaging profile to assign the numbers to.}

\item{params}{Optional further arguments to pass into API body. Should be a named list.}

\item{...}{Unused at this time.}
}
\description{
Purchase Telnyx numbers and optionally assign them to a Messaging Profile.
This leverages a combination of the Number Search API and Number Order API to find
numbers based on a series of search parameters.

Currently only configured to purchase US phone numbers.

See \href{https://developers.telnyx.com/docs/v2/numbers/quickstarts/number-search}{here} for more details.
}
