% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_send_msg_profile}
\alias{tl_send_msg_profile}
\title{Send SMS via a Number Pool attached to Telnyx Messaging Profil}
\usage{
tl_send_msg_profile(to, message, profile_id, key, params = NULL)
}
\arguments{
\item{to}{the number to send the message to}

\item{message}{the message to send}

\item{profile_id}{messaging profile id}

\item{key}{Telnyx API key}

\item{params}{optional further params to pass into the API call. Should be a named list.}
}
\description{
Send an SMS through the Number Pool Feature of Telnyx. Specifically to the \code{"https://api.telnyx.com/v2/messages/number_pool"} endpoint.
This is the recommended way of sending texts because it takes advantage of all the Messaging Profile features
Which include things like matching locale and sticky sender. You also don't need need to remember which phone number to use as the sending number.

See \href{https://developers.telnyx.com/docs/api/v2/messaging/Messaging-Profiles}{here} for more details.
}
