% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_send_text}
\alias{tl_send_text}
\title{Send text with Telnyx}
\usage{
tl_send_text(to, from, message, key, params = NULL)
}
\arguments{
\item{to}{number to send the text to}

\item{from}{number to send the text from}

\item{message}{character string of the text}

\item{key}{Telnyx API key}

\item{params}{optional further params to pass into the API call. Should be a named list.}
}
\description{
Send a text with the Telnyx API. Specifically to the \code{"https://api.telnyx.com/v2/messages"} endpoint
Note that using \code{tl_send_messaging_profile} is the preferred way of sending multiple SMS as it is more feature rich.

See \href{https://developers.telnyx.com/docs/api/v2/messaging/}{here} for more details.
}
