% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_set_call_fwd}
\alias{tl_set_call_fwd}
\title{Setup call forwarding for a given Messaging Profile}
\usage{
tl_set_call_fwd(key, fwd_number, profile_id = NULL, phone_ids = NULL, ...)
}
\arguments{
\item{key}{Required. Telnyx API key.}

\item{fwd_number}{Required. Phone number to forward to. Should be a string in the format +14156660420.}

\item{profile_id}{Messaging Profile id to update all numbers. Required if \code{phone_ids} are not set.}

\item{phone_ids}{Vector of phone ids to update. Note these are not the phone numbers, but the Telnyx assigned ID. Required if \code{profile_id} is not set.}

\item{...}{Unused at this time.}
}
\description{
Sets call forwarding up for a given Messaging Profile.

See \href{https://developers.telnyx.com/docs/api/v2/numbers/Number-Configurations}{here} for more details.
}
