% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telnyx.R
\name{tl_setup_msg_profile}
\alias{tl_setup_msg_profile}
\title{Setup Messaging Profile with numbers and Call Forwarding}
\usage{
tl_setup_msg_profile(
  api_key,
  profile_name,
  n_phone_numbers,
  profile_state,
  profile_city = NULL,
  call_fwd_number = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{api_key}{Required. Telnyx API key.}

\item{profile_name}{Required. Messaging Profile name.}

\item{n_phone_numbers}{Required. Number of phone numbers to allocate to the Messaging Service.}

\item{profile_state}{Required. Abbreviated state name to look up phone numbers. e.g. CA, OR, CO, etc.}

\item{profile_city}{Optional. City name to narrow down the search to provision phone numbers.}

\item{call_fwd_number}{Optional. Phone number to forward to. Should be a string in the format +14156660420.}

\item{verbose}{Defaults to \code{FALSE}.}

\item{...}{Optional further parameters to pass into other functions.}
}
\description{
Setup a messaging profile, purchase a set of numbers, and apply call forwarding.
This is a convenience wrapper around \code{tl_create_msg_profile}, \code{tl_order_numbers}, and \code{tl_set_call_fwd}
If you want more specific settings use those functions individually or pass in function parameters.

See \href{https://developers.telnyx.com/docs/api/v2/numbers/Number-Configurations}{here} for more details.
}
