% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw_get_messages}
\alias{tw_get_messages}
\title{Get SMS messages.}
\usage{
tw_get_messages(date = NULL, from = NULL, to = NULL,
  get_all = FALSE, verbose = FALSE, sid = Sys.getenv("TWILIO_SID"),
  token = Sys.getenv("TWILIO_TOKEN"), params = NULL)
}
\arguments{
\item{date}{the date the message was sent. Defaults to \code{NULL} which gives you all dates.}

\item{from}{the phone number the message was sent from. Defaults to \code{NULL} which gives you all sending phone numbers.}

\item{to}{the phone number the message was sent to. Defaults to \code{NULL} which gives you all recieving phone numbers.}

\item{get_all}{should we get all messages instead of the first 50? Defaults to \code{FALSE}. Caution, lots of messages will take a lot of time.}

\item{verbose}{optional for verbose output. Defaults to \code{FALSE}}

\item{sid}{Twilio account sid. Defaults to \code{Sys.getenv('TWILIO_SID')}}

\item{token}{Twilio account token. Defaults to \code{Sys.getenv('TWILIO_TOKEN')}}

\item{params}{optional futher arguments to pass into API query. Should be a named list.}
}
\description{
Get SMS messages from the  \code{"https://api.twilio.com/2010-04-01/Accounts/YOUR_SID/Messages.json"} endpoint.
This can loop through the pagination of the messages for pulling down more than 50 messages at a time.

See \href{https://www.twilio.com/docs/sms/services}{here} for more details.
}
