% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw_get_messaging_service_nums}
\alias{tw_get_messaging_service_nums}
\title{Get numbers from a Twilio Messaging Service}
\usage{

  tw_get_messaging_service_nums(service_sid = Sys.getenv("TWILIO_MESSAGING_SERVICE"),
  sid = Sys.getenv("TWILIO_SID"), token = Sys.getenv("TWILIO_TOKEN"))
}
\arguments{
\item{service_sid}{messaging service sid. Defaults to \code{Sys.getenv('TWILIO_MESSAGING_SERVICE')}.}

\item{sid}{Twilio account sid. Defaults to \code{Sys.getenv('TWILIO_SID')}}

\item{token}{Twilio account token. Defaults to \code{Sys.getenv('TWILIO_TOKEN')}}
}
\description{
Get phone numbers from a Messaging Service from the \code{"https://messaging.twilio.com/v1/Services/SERVICE_SID/PhoneNumbers"} endpoint.

See \href{https://www.twilio.com/docs/sms/services}{here} for more details.
}
