% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw_notify_twilio}
\alias{tw_notify_twilio}
\title{Send a message through the Twilio Notify Service.}
\usage{
tw_notify_twilio(message, tag,
  service_sid = Sys.getenv("TWILIO_NOTIFY_SID"),
  sid = Sys.getenv("TWILIO_SID"), token = Sys.getenv("TWILIO_TOKEN"),
  params = NULL)
}
\arguments{
\item{message}{character string of the text}

\item{tag}{character string of the Twilio Notify tag you are messaging. See \href{https://www.twilio.com/docs/notify/send-notifications}{here}}

\item{service_sid}{optional explicit decleration of \code{TWILIO_NOTIFY_SID} defaults to \code{Sys.getenv('TWILIO_NOTIFY_SERVICE'}. See \href{https://www.twilio.com/docs/notify/quickstart/sms#gather-account-information}{here}..}

\item{sid}{Twilio account sid. Defaults to \code{Sys.getenv('TWILIO_SID')}}

\item{token}{Twilio account token. Defaults to \code{Sys.getenv('TWILIO_TOKEN')}}

\item{params}{further optional vakues to send into the API call (e.g. \code{identity}). Should be a named ;ist. See \href{https://www.twilio.com/docs/notify/api/notification-resource}{here}}
}
\description{
Send a message with the Twilio Notify API. Specifically to the \code{"https://notify.twilio.com/v1/Services/YOUR_NOTIFY_TOKEN/Notifications"} endpoin.

See \href{https://www.twilio.com/docs/notify/send-notifications}{here} for more details.
Unless you are a user of the Notify API and are trying to send notifcations across device types I suggest using \code{tw_send_messaging_service}
With your favorite \code{apply}, \code{purrr:map}, or \code{for(m in messages)} equivalent.
}
