% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw_send_messaging_service}
\alias{tw_send_messaging_service}
\title{Send SMS via the Twilio Messaging service}
\usage{
tw_send_messaging_service(to, message,
  service_sid = Sys.getenv("TWILIO_MESSAGING_SERVICE"),
  sid = Sys.getenv("TWILIO_SID"), token = Sys.getenv("TWILIO_TOKEN"),
  params = NULL)
}
\arguments{
\item{to}{the number to send the message to}

\item{message}{the message to send}

\item{service_sid}{messaging service sid. Defaults to \code{Sys.getenv('TWILIO_MESSAGING_SERVICE')}.}

\item{sid}{Twilio account sid. Defaults to \code{Sys.getenv('TWILIO_SID')}}

\item{token}{Twilio account token. Defaults to \code{Sys.getenv('TWILIO_TOKEN')}}

\item{params}{further params to pass into the API call. Should be a named list.}
}
\description{
Send an SMS through the Twilio Messaging Service. Specifically to the \code{"https://api.twilio.com/2010-04-01/Accounts/YOUR_SID/Messages.json"} endpoint.
This is the recommended way of sending texts because it takes advantage of all the \href{https://www.twilio.com/copilot}{Copilot} features
Which include things like matching locale and sticky sender. You also don't need need to remember which phone number to use as the sending number.

See \href{https://www.twilio.com/docs/sms/services}{here} for more details.
}
