% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw_send_text}
\alias{tw_send_text}
\title{Send text with Twilio}
\usage{
tw_send_text(to, from, message, sid = Sys.getenv("TWILIO_SID"),
  token = Sys.getenv("TWILIO_TOKEN"), params = NULL)
}
\arguments{
\item{to}{number to send the text to}

\item{from}{number to send the text from}

\item{message}{character string of the text}

\item{sid}{Twilio account sid. Defaults to \code{Sys.getenv('TWILIO_SID')}}

\item{token}{Twilio account token. Defaults to \code{Sys.getenv('TWILIO_TOKEN')}}

\item{params}{optional further params to pass into the API call. Should be a named list.}
}
\description{
Send a text with the Twilio API. Specifically to the \code{"https://api.twilio.com/2010-04-01/Accounts/YOUR_SID/Messages.json"} endpoint
Works similar \code{twilio::tw_send_message()} only we are explicit that this is a text message, and allow the \code{sid} and \code{token} to be passed in as parameters
Note that using \code{tw_send_messaging_service} is the preferred way of sending multiple SMS as it is more feature rich.

See \href{https://www.twilio.com/docs/sms/send-messages}{here} for more details.
}
